+++ test-utils/uri/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "uri"
+edition.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+tempfile = { workspace = true }
