+++ test-utils/uri/src/tempdir.rs
@@ -1,22 +1,7 @@
 use anyhow::{anyhow, Result};
 use tempfile::TempDir;
 
-pub trait TestArrayUri {
-    fn base_dir(&self) -> Result<String>;
-    fn with_paths(&self, paths: &[&str]) -> Result<String>;
-    fn close(self) -> Result<()>;
-
-    fn with_path(&self, path: &str) -> Result<String> {
-        self.with_paths(&[path])
-    }
-}
-
-pub fn get_uri_generator() -> Result<impl TestArrayUri> {
-    // TODO: Eventually this will check an environment variable to decide
-    // whether we should return a TestDirectory or a new struct called something
-    // like TestRestServer to run our test suite against the cloud service.
-    TestDirectory::new()
-}
+use super::TestArrayUri;
 
 pub struct TestDirectory {
     base_dir: TempDir,
