+++ tiledb/api/Cargo.toml
@@ -21,13 +21,13 @@ tiledb-common = { workspace = true }
 tiledb-proc-macro = { workspace = true }
 tiledb-serde = { workspace = true, optional = true, features = ["serde"] }
 tiledb-sys = { workspace = true }
-tiledb-test-utils = { workspace = true, optional = true }
 
 [dev-dependencies]
+cells = { workspace = true }
 num-traits = { workspace = true }
 proptest = { workspace = true }
 tiledb-serde = { workspace = true, features = ["proptest-strategies", "option-subset", "serde"] }
-tiledb-test-utils = { workspace = true }
+uri = { workspace = true }
 
 [build-dependencies]
 pkg-config = { workspace = true }
