--- tiledb/test-utils/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "tiledb-test-utils"
-edition.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[dependencies]
-anyhow = { workspace = true, optional = true }
-nix = { version = "0", features = ["signal"], optional = true }
-paste = { workspace = true, optional = true }
-tempfile = { workspace = true, optional = true }
-
-[features]
-default = ["uri"]
-signal = ["dep:nix"]
-uri = ["dep:anyhow", "dep:tempfile"]
